requirements.txt
ShopifyAPI
python-dotenv